requirements.txt
keybert==0.9.0
sentence-transformers==5.0.0
weaviate-client==4.16.0
torch==2.7.1
numpy==2.3.1
pillow==11.3.0
tqdm==4.67.1
torchvision==0.22.1
xformers==0.0.31.post1
git+https://github.com/openai/CLIP.git
pandas==2.3.1